exercise 4.21

+++ part_4/blog/controllers/blogs.js
@@ -45,6 +45,20 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(400).json({ error: "blog not found" });
+  }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(400).json({ error: "user is not owner of blog" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
